Fixed an error in extension processing.
If a file has no extension, but is in a directory whose name includes a '.', the fully-qualified filename after the period was read as the extension.
Added a unit test to cover this case, and fixed it by adding a search for the file system directory separator (os.sep) into the extension processing.