Detector stub for extensions.
I'm trying a design change.  I only want to walk a given tree once.  I want to be able to do all the processing that I need to do, on that tree,
in a modular enough way that I can call the tree walking with functions and variables that carry out whatever data the functions need.  Maybe this calls for an
object-oriented design, where I pass in the accumulating calculation?