Filtering at the filename level
You don't have to iterate through every filename in a directory tree, with the goal of saving time.  If you pass in a boolean-returning function that takes the filename and returns true if it should be filtered out, then you can selectively ignore files based on whatever criteria you can get out of the filename.